moved instructions-specific stuff to its component file
